requirements.txt
arrow==1.4.0
asgiref==3.10.0
binaryornot==0.4.4
certifi==2025.10.5
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
cookiecutter==2.6.0
Django==5.2.7
# django-ckeditor==6.7.0  # Commented out to prevent CKEditor conflicts
django-tailwind==4.2.0
djangorestframework==3.16.1
psycopg2-binary==2.9.11
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
six==1.17.0
imapclient==2.3.1
dnspython==2.4.2
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
